+++ src/main/java/com/balybus/galaxy/chat/controller/ChatSocketController.java
@@ -1,6 +1,6 @@
 package com.balybus.galaxy.chat.controller;
 
-import com.balybus.galaxy.chat.domain.tblChatMsg.TblChatMsg;
+import com.balybus.galaxy.global.domain.tblChatMsg.TblChatMsg;
 import com.balybus.galaxy.chat.dto.ChatMsgRequestDto;
 import com.balybus.galaxy.chat.dto.ChatMsgResponseDto;
 import com.balybus.galaxy.chat.service.ChatServiceImpl;
